+++ lib/glimmer/excluded_keyword_error.rb
@@ -1,0 +1,5 @@
+module Glimmer
+  # Represents Glimmer errors that occur due to excluded keywords passing through method_missing
+  class ExcludedKeywordError < RuntimeError
+  end
+end