+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.9.3
 
 - Add infinite loop detection support to avoid infinite loops (breaks after 10 repetitions)
+- ExcludedKeywordError with no logging for excluded keywords such as `to_a` and `to_h`
 
 ## 0.9.2